splitting entities reducer for supporting more reducers on EE
