+++ app/client/src/pages/Editor/APIEditor/index.tsx
@@ -10,7 +10,7 @@ import type { RouteComponentProps } from "react-router";
 import type {
   ActionData,
   ActionDataState,
-} from "reducers/entityReducers/actionsReducer";
+} from "@appsmith/reducers/entityReducers/actionsReducer";
 import _ from "lodash";
 import { getCurrentApplication } from "@appsmith/selectors/applicationSelectors";
 import AnalyticsUtil from "utils/AnalyticsUtil";
