+++ app/client/src/sagas/ActionSagas.ts
@@ -64,7 +64,7 @@ import {
 import type {
   ActionData,
   ActionDataState,
-} from "reducers/entityReducers/actionsReducer";
+} from "@appsmith/reducers/entityReducers/actionsReducer";
 import {
   getAction,
   getActions,
