+++ app/client/src/sagas/SnipingModeSagas.ts
@@ -2,7 +2,7 @@ import { all, call, put, select, takeLeading } from "redux-saga/effects";
 import type { ReduxAction } from "@appsmith/constants/ReduxActionConstants";
 import { ReduxActionTypes } from "@appsmith/constants/ReduxActionConstants";
 import { snipingModeBindToSelector } from "selectors/editorSelectors";
-import type { ActionData } from "reducers/entityReducers/actionsReducer";
+import type { ActionData } from "@appsmith/reducers/entityReducers/actionsReducer";
 import { getCanvasWidgets } from "@appsmith/selectors/entitiesSelector";
 import {
   batchUpdateWidgetDynamicProperty,
