+++ app/client/src/sagas/WidgetLoadingSaga.ts
@@ -9,7 +9,7 @@ import { getActions } from "@appsmith/selectors/entitiesSelector";
 import type {
   ActionData,
   ActionDataState,
-} from "reducers/entityReducers/actionsReducer";
+} from "@appsmith/reducers/entityReducers/actionsReducer";
 import type { ReduxAction } from "@appsmith/constants/ReduxActionConstants";
 import {
   ReduxActionErrorTypes,
